+++ app/(route)/my-project/_components/RecordingAnimation.tsx
