+++ app/(route)/my-project/_components/RemoveContentBtn.tsx
@@ -1,5 +1,5 @@
-import { MinusIcon } from "../../../../components/Icons";
-import { BtnProps } from "../../_types/project_types";
+import { MinusIcon } from "../../../components/Icons";
+import { BtnProps } from "../_types/project_types";
 
 export default function RemoveContentBtn({ onClick }: BtnProps) {
   return (
