--- app/(route)/my-project/new-project/record/_components/Icons.tsx
@@ -1,75 +1,0 @@
-// 녹음 중 아이콘
-export const RecordingIcon = () => {
-  return (
-    <svg
-      width="100"
-      height="100"
-      viewBox="0 0 100 100"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M29.6875 75.0002V25.0002H35.9375V75.0002H29.6875ZM46.875 91.6668V8.3335H53.125V91.6668H46.875ZM12.5 58.3335V41.6668H18.75V58.3335H12.5ZM64.0625 75.0002V25.0002H70.3125V75.0002H64.0625ZM81.25 58.3335V41.6668H87.5V58.3335H81.25Z"
-        fill="#6B71FF"
-      />
-    </svg>
-  );
-};
-
-//  재생 아이콘
-export const PlayIcon = () => {
-  return (
-    <svg
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M9.5 16.5L16.5 12L9.5 7.5V16.5ZM12 22C10.6167 22 9.31667 21.7375 8.1 21.2125C6.88333 20.6875 5.825 19.975 4.925 19.075C4.025 18.175 3.3125 17.1167 2.7875 15.9C2.2625 14.6833 2 13.3833 2 12C2 10.6167 2.2625 9.31667 2.7875 8.1C3.3125 6.88333 4.025 5.825 4.925 4.925C5.825 4.025 6.88333 3.3125 8.1 2.7875C9.31667 2.2625 10.6167 2 12 2C13.3833 2 14.6833 2.2625 15.9 2.7875C17.1167 3.3125 18.175 4.025 19.075 4.925C19.975 5.825 20.6875 6.88333 21.2125 8.1C21.7375 9.31667 22 10.6167 22 12C22 13.3833 21.7375 14.6833 21.2125 15.9C20.6875 17.1167 19.975 18.175 19.075 19.075C18.175 19.975 17.1167 20.6875 15.9 21.2125C14.6833 21.7375 13.3833 22 12 22ZM12 20C14.2333 20 16.125 19.225 17.675 17.675C19.225 16.125 20 14.2333 20 12C20 9.76667 19.225 7.875 17.675 6.325C16.125 4.775 14.2333 4 12 4C9.76667 4 7.875 4.775 6.325 6.325C4.775 7.875 4 9.76667 4 12C4 14.2333 4.775 16.125 6.325 17.675C7.875 19.225 9.76667 20 12 20Z"
-        fill="#6B71FF"
-      />
-    </svg>
-  );
-};
-
-// 마이크 아이콘
-export const MikeIcon = () => {
-  return (
-    <svg
-      width="100"
-      height="100"
-      viewBox="0 0 100 100"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M32.6895 100C31.7992 100 31.0589 99.6988 30.4687 99.096C29.8784 98.4939 29.5833 97.7474 29.5833 96.8564C29.5833 95.9661 29.8846 95.2259 30.4874 94.6356C31.0895 94.0453 31.836 93.7502 32.727 93.7502C33.6173 93.7502 34.3576 94.0516 34.9478 94.6543C35.5381 95.2564 35.8333 96.0029 35.8333 96.8939C35.8333 97.7842 35.5319 98.5245 34.9291 99.1147C34.327 99.705 33.5805 100 32.6895 100ZM49.9999 100C49.1437 100 48.4263 99.6877 47.8478 99.0627C47.2687 98.4377 46.9791 97.7085 46.9791 96.8752C46.9791 96.0418 47.2732 95.3127 47.8614 94.6877C48.4503 94.0627 49.1794 93.7502 50.0489 93.7502C50.919 93.7502 51.6666 94.0516 52.2916 94.6543C52.9166 95.2564 53.2291 96.0029 53.2291 96.8939C53.2291 97.7842 52.9197 98.5245 52.301 99.1147C51.6815 99.705 50.9145 100 49.9999 100ZM67.377 100C66.4867 100 65.7464 99.6988 65.1562 99.096C64.5659 98.4939 64.2708 97.7474 64.2708 96.8564C64.2708 95.9661 64.5721 95.2259 65.1749 94.6356C65.777 94.0453 66.5235 93.7502 67.4145 93.7502C68.3048 93.7502 69.0451 94.0516 69.6353 94.6543C70.2256 95.2564 70.5208 96.0029 70.5208 96.8939C70.5208 97.7842 70.2194 98.5245 69.6166 99.1147C69.0145 99.705 68.268 100 67.377 100ZM49.9999 55.9377C47.0138 55.9377 44.5138 54.864 42.4999 52.7168C40.486 50.5703 39.4791 47.9634 39.4791 44.896V18.7502C39.4791 15.8564 40.5013 13.397 42.5458 11.372C44.5902 9.34634 47.0728 8.3335 49.9937 8.3335C52.9145 8.3335 55.3992 9.34634 57.4478 11.372C59.4964 13.397 60.5208 15.8564 60.5208 18.7502V44.896C60.5208 47.9634 59.5138 50.5703 57.4999 52.7168C55.486 54.864 52.986 55.9377 49.9999 55.9377ZM46.8749 87.5002V73.3335C39.5138 72.5696 33.3333 69.4793 28.3333 64.0627C23.3333 58.646 20.8333 52.2571 20.8333 44.896H27.0833C27.0833 51.2154 29.3155 56.5279 33.7801 60.8335C38.2447 65.139 43.644 67.2918 49.978 67.2918C56.3121 67.2918 61.7187 65.139 66.1978 60.8335C70.677 56.5279 72.9166 51.2154 72.9166 44.896H79.1666C79.1666 52.2571 76.6666 58.646 71.6666 64.0627C66.6666 69.4793 60.486 72.5696 53.1249 73.3335V87.5002H46.8749Z"
-        fill="#6B71FF"
-      />
-    </svg>
-  );
-};
-
-// 녹음 버튼 틀
-export const RecordIcon = () => {
-  return (
-    <svg
-      width="60"
-      height="60"
-      viewBox="0 0 60 60"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <circle
-        cx="30"
-        cy="30"
-        r="27.5"
-        fill="#F7F7F7"
-        stroke="#737373"
-        stroke-width="5"
-      />
-    </svg>
-  );
-};
